refactor(parser): remove standalone parser (#764)
