--- src/handler/src/types/broker.ts
@@ -1,9 +1,0 @@
-import { Token } from 'lexure';
-import { Message } from '@spectacles/types';
-
-export interface BrokerParserOutput {
-	command: Token;
-	arguments: Record<string, unknown>;
-	tokens: Token[];
-	message: Message;
-}
