--- src/parser/package.json
@@ -1,28 +1,0 @@
-{
-	"name": "@yuudachi/parser",
-	"version": "2.0.0",
-	"description": "",
-	"main": "dist/index.js",
-	"scripts": {
-		"build": "rimraf dist && tsc",
-		"start": "yarn node dist/bin/index.js",
-		"dev": "yarn run build && env-cmd yarn run start"
-	},
-	"keywords": [],
-	"author": "iCrawl <icrawltogo@outlook.com>",
-	"license": "UNLICENSED",
-	"private": true,
-	"dependencies": {
-		"@spectacles/brokers": "^0.8.0",
-		"lexure": "^0.14.1",
-		"postgres": "^2.0.0-beta.0",
-		"tslib": "^2.0.0"
-	},
-	"devDependencies": {
-		"@spectacles/types": "^0.2.1",
-		"@types/node": "^14.0.27",
-		"env-cmd": "^10.1.0",
-		"rimraf": "^3.0.2",
-		"typescript": "^3.9.7"
-	}
-}
