--- src/parser/tsconfig.eslint.json
@@ -1,9 +1,0 @@
-{
-	"extends": "./tsconfig.json",
-	"include": [
-		"**/*.ts",
-		"**/*.test.ts",
-		"**/*.spec.ts",
-	],
-	"exclude": []
-}
