--- src/parser/tsconfig.json
@@ -1,12 +1,0 @@
-{
-	"extends": "../../tsconfig.base.json",
-	"compilerOptions": {
-		"module": "CommonJS",
-		"sourceRoot": "./",
-		"rootDir": ".",
-		"outDir": "dist"
-	},
-	"include": [
-		"./bin"
-	]
-}
